plugins/synvo_ai_mail/requirements.txt
azure-identity>=1.14.0
azure-core>=1.29.0
msal>=1.24.0
msal-extensions>=1.0.0
msgraph-sdk>=1.0.0
mail>=2.1.0
